moved styled-components of helpers to helpers directory
