+++ src/components/Helpers/style/LikedHeart.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { colors } from "../variables";
+import { colors } from "../../../styles/variables";
 
 const { likedPrimary, likedSecondary } = colors;
 export const NotLikedHeart = styled.div`
