--- src/styles/Helpers/index.js
@@ -1,11 +1,0 @@
-import { LikedHeart, NotLikedHeart } from "./LikedHeart";
-import UserLike from "./UserLike";
-
-import ImageWrapper from "./ImageWrapperStyle";
-
-export const ImageWrapperStyle = ImageWrapper;
-
-// Like component styles
-export const UserLikeStyle = UserLike;
-export const LikedHeartStyle = LikedHeart;
-export const NotLikedHeartStyle = NotLikedHeart;
